refactor: page별로 파일 분리 - NewProject - 프로젝트 생성 페이지 파일 분리
